add version locking and checking to ripgrep

+++ packages/ripgrep.sh
@@ -1,8 +1,20 @@
 #!/bin/bash
 
-git clone https://github.com/BurntSushi/ripgrep.git
-pushd ripgrep
-  cargo build --release
-  sudo mv target/release/rg /usr/local/bin
+VERSION=0.10.0
+
+if [[ `rg --version | head -n1` == "ripgrep $VERSION"* ]]; then
+    echo "Looks like ripgrep $VERSION is already installed, skipping..."
+    exit
+fi
+
+pushd $HOME/workspace
+  mkdir ripgrep
+  pushd ripgrep
+    wget https://github.com/BurntSushi/ripgrep/archive/$VERSION.zip
+    unzip $VERSION.zip
+    cd ripgrep-$VERSION
+    cargo build --release
+    sudo mv target/release/rg /usr/local/bin
+  popd
+  rm -rf ripgrep
 popd
-rm -rf ripgrep
